requirements.txt
discord.py==0.11.0
requests==2.10.0
beautifulsoup4==4.5.1
